+++ src/java/src/test/java/com/dnanexus/DXSearchTest.java
@@ -272,6 +272,77 @@ public class DXSearchTest {
         Assert.assertEquals("aworkflow", workflowResult.describe().getName());
     }
 
+    /**
+     * Tests retrieving Describe output with findDataObjects.
+     */
+    @Test
+    public void testFindDataObjectsWithDescribe() {
+        DXRecord.newRecord().setProject(testProject).setName("record1")
+                .putProperty("sampleId", "1234").build();
+        DXFile.newFile().setProject(testProject).setName("file1").putProperty("sampleId", "2345")
+                .build();
+        DXGTable.newGTable(
+                ImmutableList.of(ColumnSpecification.getInstance("num_goats", ColumnType.INT16)))
+                .setProject(testProject).setName("gtable1").build();
+        DXApplet.newApplet().setProject(testProject).setName("applet1")
+                .setRunSpecification(RunSpecification.newRunSpec("bash", "").build()).build();
+        DXWorkflow.newWorkflow().setProject(testProject).setName("workflow1").build();
+
+        DXRecord recordResult =
+                Iterables.getOnlyElement(DXSearch.findDataObjects().withClassRecord()
+                        .inProject(testProject).nameMatchesExactly("record1")
+                        .includeDescribeOutput(DXDataObject.DescribeOptions.get().withProperties())
+                        .execute().asList());
+        Assert.assertEquals(recordResult.getCachedDescribe().getName(), "record1");
+        // Called includeDescribeOutput with properties: true so properties should be returned
+        Assert.assertEquals(recordResult.getCachedDescribe().getProperties().get("sampleId"),
+                "1234");
+
+        DXGTable gtableResult =
+                Iterables.getOnlyElement(DXSearch.findDataObjects().withClassGTable()
+                        .inProject(testProject).nameMatchesExactly("gtable1")
+                        .includeDescribeOutput().execute().asList());
+        Assert.assertEquals(gtableResult.getCachedDescribe().getName(), "gtable1");
+        // Called includeDescribeOutput with default settings so properties should NOT be returned
+        try {
+            gtableResult.getCachedDescribe().getProperties();
+            Assert.fail("Expected NullPointerException to be thrown because properties should not have been returned");
+        } catch (NullPointerException e) {
+            // Expected
+        }
+
+        Assert.assertEquals(
+                Iterables
+                        .getOnlyElement(
+                                DXSearch.findDataObjects().inProject(testProject)
+                                        .nameMatchesExactly("file1").includeDescribeOutput()
+                                        .execute().asList()).getCachedDescribe().getName(), "file1");
+        Assert.assertEquals(
+                Iterables
+                        .getOnlyElement(
+                                DXSearch.findDataObjects().inProject(testProject)
+                                        .nameMatchesExactly("applet1").includeDescribeOutput()
+                                        .execute().asList()).getCachedDescribe().getName(),
+                "applet1");
+        Assert.assertEquals(
+                Iterables
+                        .getOnlyElement(
+                                DXSearch.findDataObjects().inProject(testProject)
+                                        .nameMatchesExactly("workflow1").includeDescribeOutput()
+                                        .execute().asList()).getCachedDescribe().getName(),
+                "workflow1");
+
+        DXRecord findWithoutDescribe =
+                Iterables.getOnlyElement(DXSearch.findDataObjects().inProject(testProject)
+                        .nameMatchesExactly("record1").withClassRecord().execute().asList());
+        try {
+            findWithoutDescribe.getCachedDescribe();
+            Assert.fail("Expected IllegalStateException to be thrown");
+        } catch (IllegalStateException e) {
+            // Expected
+        }
+    }
+
     @Test
     public void testFindJobs() {
         @SuppressWarnings("unused")
@@ -329,6 +400,15 @@ public class DXSearchTest {
                 DXJSON.parseJson("{\"level\": \"ADMINISTER\"}"),
                 mapper.valueToTree(DXSearch.findDataObjects()
                         .withMinimumAccessLevel(AccessLevel.ADMINISTER).buildRequestHash()));
+        Assert.assertEquals(
+                DXJSON.parseJson("{\"describe\": true}"),
+                mapper.valueToTree(DXSearch.findDataObjects().includeDescribeOutput()
+                        .buildRequestHash()));
+        Assert.assertEquals(
+                DXJSON.parseJson("{\"describe\": {\"properties\": true}}"),
+                mapper.valueToTree(DXSearch.findDataObjects()
+                        .includeDescribeOutput(DXDataObject.DescribeOptions.get().withProperties())
+                        .buildRequestHash()));
 
         try {
             DXSearch.findDataObjects().inProject(DXProject.getInstance("project-0000"))